vk objects code polishing
