+++ source/GlobalUsings.cs
@@ -2,7 +2,6 @@
 global using System.Collections.Generic;
 global using System.IO;
 global using System.Linq;
-global using System.Threading;
-global using System.Threading.Tasks;
 
+global using Silk.NET.Core.Native;
 global using Silk.NET.Vulkan;
