--- source/api/vulkan/VkVertexBuffer.cs
@@ -1,7 +1,0 @@
-﻿/*namespace Vulkanoid.Vulkan;
-
-public sealed class VkVertexBuffer : VkDeviceBuffer<Vertex>
-{
-    public VkVertexBuffer(VkDevice device, VkCommandPool commandPool, Vertex[] vertices)
-        : base(device, commandPool, vertices, BufferUsageFlags.VertexBufferBit) { }
-}*/
