+++ source/api/vulkan/device/VkDevice.CommandBuffer.cs
@@ -1,0 +1,6 @@
+namespace Vulkanoid.Vulkan;
+
+public sealed partial class VkDevice : GraphicsDevice
+{
+    public void FreeCommandBuffer(VkCommandBuffer commandBuffer) => vk.FreeCommandBuffers(handle, commandBuffer.commandPool, 1u, commandBuffer);
+}
