+++ source/api/vulkan/device/VkDevice.RenderPass.cs
@@ -1,0 +1,97 @@
+namespace Vulkanoid.Vulkan;
+
+public sealed partial class VkDevice : GraphicsDevice
+{
+    public VkRenderPass CreateRenderPass(Format format, SampleCountFlags sampleCount)
+    {
+        var colorAttachment = new AttachmentDescription
+        {
+            Format = format,
+            Samples = sampleCount,
+            LoadOp = AttachmentLoadOp.Clear,
+            StoreOp = AttachmentStoreOp.Store,
+            StencilLoadOp = AttachmentLoadOp.DontCare,
+            StencilStoreOp = AttachmentStoreOp.DontCare,
+            InitialLayout = ImageLayout.Undefined,
+            FinalLayout = ImageLayout.ColorAttachmentOptimal
+        };
+
+        var colorAttachmentResolve = new AttachmentDescription
+        {
+            Format = format,
+            Samples = SampleCountFlags.Count1Bit,
+            LoadOp = AttachmentLoadOp.DontCare,
+            StoreOp = AttachmentStoreOp.Store,
+            StencilLoadOp = AttachmentLoadOp.DontCare,
+            StencilStoreOp = AttachmentStoreOp.DontCare,
+            InitialLayout = ImageLayout.Undefined,
+            FinalLayout = ImageLayout.PresentSrcKhr
+        };
+
+        var depthAttachment = new AttachmentDescription
+        {
+            Format = physicalDevice.DepthFormat,
+            Samples = sampleCount,
+            LoadOp = AttachmentLoadOp.Clear,
+            StoreOp = AttachmentStoreOp.DontCare,
+            StencilLoadOp = AttachmentLoadOp.DontCare,
+            StencilStoreOp = AttachmentStoreOp.DontCare,
+            InitialLayout = ImageLayout.Undefined,
+            FinalLayout = ImageLayout.DepthStencilAttachmentOptimal
+        };
+
+        var colorAttachmentRef = new AttachmentReference
+        {
+            Attachment = 0,
+            Layout = ImageLayout.ColorAttachmentOptimal
+        };
+
+        var depthAttachmentRef = new AttachmentReference
+        {
+            Attachment = 1,
+            Layout = ImageLayout.DepthStencilAttachmentOptimal
+        };
+
+        var colorAttachmentResolveRef = new AttachmentReference
+        {
+            Attachment = 2,
+            Layout = ImageLayout.ColorAttachmentOptimal
+        };
+
+        unsafe
+        {
+            var subpass = new SubpassDescription
+            {
+                PipelineBindPoint = PipelineBindPoint.Graphics,
+                ColorAttachmentCount = 1,
+                PColorAttachments = &colorAttachmentRef,
+                PDepthStencilAttachment = &depthAttachmentRef,
+                PResolveAttachments = &colorAttachmentResolveRef
+            };
+
+            var dependency = new SubpassDependency
+            {
+                SrcSubpass = Vk.SubpassExternal,
+                DstSubpass = 0,
+                SrcStageMask = PipelineStageFlags.ColorAttachmentOutputBit | PipelineStageFlags.EarlyFragmentTestsBit,
+                SrcAccessMask = 0,
+                DstStageMask = PipelineStageFlags.ColorAttachmentOutputBit | PipelineStageFlags.EarlyFragmentTestsBit,
+                DstAccessMask = AccessFlags.ColorAttachmentReadBit | AccessFlags.ColorAttachmentWriteBit | AccessFlags.DepthStencilAttachmentWriteBit
+            };
+
+            var attachments = stackalloc AttachmentDescription[3] { colorAttachment, depthAttachment, colorAttachmentResolve };
+
+            var renderPassInfo = new RenderPassCreateInfo(
+                attachmentCount: 3,
+                pAttachments: attachments,
+                subpassCount: 1,
+                pSubpasses: &subpass,
+                dependencyCount: 1,
+                pDependencies: &dependency);
+
+            vk.CreateRenderPass(handle, renderPassInfo, null, out var renderPassHandle);
+
+            return new VkRenderPass(renderPassHandle, this);
+        }
+    }
+}
